Conflicts:
	Android/res/values/strings.xml
	Android/src/org/droidplanner/android/activities/EditorActivity.java
	Android/src/org/droidplanner/android/fragments/EditorMapFragment.java
	Android/src/org/droidplanner/android/maps/providers/google_map/GoogleMapFragment.java
	Android/src/org/droidplanner/android/maps/providers/osm/OSMapFragment.java

+++ Core/src/org/droidplanner/core/drone/variables/Navigation.java
@@ -14,11 +14,10 @@ public class Navigation extends DroneVariable {
 		super(myDrone);
 	}
 
-	public void setNavPitchRollYaw(float nav_pitch, float nav_roll,
-			short nav_bearing) {
-		this.nav_pitch = (double) nav_pitch;
-		this.nav_roll = (double) nav_roll;
-		this.nav_bearing = (double) nav_bearing;
+	public void setNavPitchRollYaw(float nav_pitch, float nav_roll, short nav_bearing) {
+		this.nav_pitch = nav_pitch;
+		this.nav_roll = nav_roll;
+		this.nav_bearing = nav_bearing;
 		myDrone.events.notifyDroneEvent(DroneEventsType.NAVIGATION);
 	}
